Docker: Documentation updates
