+++ docs/source/autodoc/crateweb/_index.rst
@@ -69,6 +69,7 @@ crate_anon/crateweb
     consent/migrations/0014_contactrequest_rdbm_to_contact_pt.py.rst
     consent/migrations/0015_auto_20190104_1255.py.rst
     consent/migrations/0016_auto_20190212_1118.py.rst
+    consent/migrations/0017_replace_nullbooleanfield.py.rst
     consent/migrations/__init__.py.rst
     consent/models.py.rst
     consent/storage.py.rst
@@ -131,6 +132,7 @@ crate_anon/crateweb
     core/context_processors.py.rst
     core/management/__init__.py.rst
     core/management/commands/__init__.py.rst
+    core/management/commands/ensuresuperuser.py.rst
     core/management/commands/runcpserver.py.rst
     core/middleware.py.rst
     core/utils.py.rst
@@ -242,8 +244,6 @@ crate_anon/crateweb
     specimen_archives/tree/snippets/results_table_nlp.mako.rst
     specimen_archives/tree/snippets/single_nlp_page.mako.rst
     specimen_archives/tree/snippets/subtree.mako.rst
-    specimen_secret_local_settings/__init__.py.rst
-    specimen_secret_local_settings/crateweb_local_settings.py.rst
     static/base.css.rst
     static/collapse.js.rst
     static/extra_styles.css.rst
