+++ docs/source/autodoc/crateweb/core/management/commands/ensuresuperuser.py.rst
@@ -1,29 +1,29 @@
-#!/usr/bin/env python
+.. docs/source/autodoc/crateweb/core/management/commands/ensuresuperuser.py.rst
+        
+.. THIS FILE IS AUTOMATICALLY GENERATED. DO NOT EDIT.
 
-"""
-crate_anon/crateweb/specimen_secret_local_settings/__init__.py
 
-===============================================================================
-
-    Copyright (C) 2015-2021 Rudolf Cardinal (rudolf@pobox.com).
-
+..  Copyright © 2015-2021 Rudolf Cardinal (rudolf@pobox.com).
+    .
     This file is part of CRATE.
-
+    .
     CRATE is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
     the Free Software Foundation, either version 3 of the License, or
     (at your option) any later version.
-
+    .
     CRATE is distributed in the hope that it will be useful,
     but WITHOUT ANY WARRANTY; without even the implied warranty of
     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
     GNU General Public License for more details.
-
+    .
     You should have received a copy of the GNU General Public License
     along with CRATE. If not, see <https://www.gnu.org/licenses/>.
 
-===============================================================================
 
-The mere existence of this file makes Python treat the directory as a package.
 
-"""
+crate_anon.crateweb.core.management.commands.ensuresuperuser
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: crate_anon.crateweb.core.management.commands.ensuresuperuser
+    :members:
