+++ docs/source/autodoc/linkage/test_hash_speed.py.rst
@@ -1,0 +1,29 @@
+.. docs/source/autodoc/linkage/test_hash_speed.py.rst
+        
+.. THIS FILE IS AUTOMATICALLY GENERATED. DO NOT EDIT.
+
+
+..  Copyright © 2015-2021 Rudolf Cardinal (rudolf@pobox.com).
+    .
+    This file is part of CRATE.
+    .
+    CRATE is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    .
+    CRATE is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU General Public License for more details.
+    .
+    You should have received a copy of the GNU General Public License
+    along with CRATE. If not, see <https://www.gnu.org/licenses/>.
+
+
+
+crate_anon.linkage.test_hash_speed
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: crate_anon.linkage.test_hash_speed
+    :members:
