+++ setup.cfg
@@ -3,7 +3,6 @@ linters=pycodestyle,pyflakes
 exclude=
     migrations,
     compiled_nlp_classes,
-    crate_anon/crateweb/specimen_secret_local_settings/crateweb_local_settings.py
 
 max-line-length=80
